% Generated by roxygen2 (4.0.1): do not edit by hand
\name{P_ibd}
\alias{P_ibd}
\title{computes the prob of identity by descent of a randomly chosen pair of gene copies after one generation}
\usage{
P_ibd(Nf, Nm, nf, y, d, r, s)
}
\arguments{
\item{Nf}{number of females that spawn exclusively in the wild}

\item{Nm}{number of males that spawn exclusively in the wild}

\item{nf}{number of females taken into the hatchery to be sub-spawned.}

\item{y}{fraction of female eggs that are taken for subspawning.}

\item{d}{number of males spawned with one female in the hathery.}

\item{r}{the fraction of "male virility" that remains after a male has been squeezed at the hatchery.}

\item{s}{number of adult returners expected from each hatchery female relative to 1 return from a wild female.}
}
\value{
Returns a double.  Actually, it will return a vector of doubles if you want to do it for multiple values.
}
\description{
This function will compute the prob of IBD after one model.  See the vignette for more of
an explanation of things.
}

